added lots

+++ src/main/java/com/example/demo/repository/MovieRepository.java
@@ -3,5 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.model.Movie;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovieRepository extends JpaRepository<Movie,Long> {
+
+public interface MovieRepository extends JpaRepository<Movie,Integer> {
 }
